kbo_attendance_prediction/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
streamlit
plotly
ollama
faiss-cpu
nbformat
nbconvert
python-pptx
